feat(ir_remote): Added driver logic for IR remote for LAUNCHPAD
The robot will decode a start signal sent at each start. The IR
transmitter will send a signal to the IR receiver on the board.
The driver will check the signal by measuring hte time between pulses
on the GPIO connected to the IR receiver and tracking the falling edges.